refactor board logic into hook
